Finish five new full-stack forms; fleet-index + generator-drift touch-ups
Completes the parallel-agent builds for:
- hernia-diagnostic-evaluation
- cataract-diagnostic-evaluation
- knee-replacement-surgery-evaluation
- hip-replacement-surgery-evaluation
- health-screening-questionnaire

This commit:
- Adds the missing typespec/{index,AGENTS,CLAUDE,README}.md placeholder
  quartet to hip-replacement-surgery-evaluation (the only one of the five
  the scaffold's typespec/ directory was left incomplete for), fixing the
  last bin/test-form failures.
- Commits health-screening-questionnaire's, cataract-diagnostic-evaluation's,
  and hernia-diagnostic-evaluation's remaining uncommitted work (the other
  two forms' build agents had already committed everything themselves).
- Adds all five forms to forms/AGENTS.md's alphabetical index and
  regenerates forms.tsv (355 forms, up from 350).
- Picks up pre-existing back-end-with-loco-setup drift on three unrelated
  forms (pre-anaesthesia-assessment, pre-operative-assessment-by-clinician,
  pre-operative-assessment-by-patient) as an incidental side effect of the
  repo-wide generate-back-end-with-loco-setup.py script the build agents
  ran -- verified as a real, beneficial fix (missing columns from an
  earlier migration were never reflected in the setup script), not
  something introduced by this task.

Verified for all five new forms: bin/test-form, bin/test-sql-apply,
bin/test-examples-conformance, bin/lily-html-refactor --check,
bin/lily-svelte-refactor --check -- all pass/clean.

+++ forms/cataract-diagnostic-evaluation/back-end-with-loco/src/cataract_diagnostic_evaluation/models/_entities/cataract_diagnostic_evaluation_grades.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub computed_locs_iii_severity_right: String,
     pub computed_locs_iii_severity_left: String,
@@ -22,7 +22,7 @@ pub struct Model {
     pub signed_by_name: String,
     pub signed_at: Option<DateTimeWithTimeZone>,
     pub graded_at: DateTimeWithTimeZone,
-    pub cataract_diagnostic_evaluation_id: i32,
+    pub cataract_diagnostic_evaluation_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
